Remove obsolete usage of transport address translation
